Moved GridObjects into it's own folder (because there's a lot of them now). And added a Shelf and a Wall
