--- src/main/java/worldofzuul/PresentationLayer/GridObject.java
@@ -1,3 +1,0 @@
-package worldofzuul.PresentationLayer;
-
-public abstract class GridObject { }
