+++ src/main/java/worldofzuul/PresentationLayer/GridObjects/GridObject.java
@@ -1,0 +1,3 @@
+package worldofzuul.PresentationLayer.GridObjects;
+
+public abstract class GridObject { }
