+++ src/main/java/worldofzuul/PresentationLayer/GridObjects/Wall.java
@@ -1,0 +1,3 @@
+package worldofzuul.PresentationLayer.GridObjects;
+
+public class Wall extends GridObject { }
