+++ src/main/java/worldofzuul/PresentationLayer/GridObjects/Warp.java
@@ -1,0 +1,28 @@
+package worldofzuul.PresentationLayer.GridObjects;
+
+import worldofzuul.PresentationLayer.Grid;
+import worldofzuul.PresentationLayer.Position;
+
+public class Warp extends GridObject {
+
+    private Grid grid;
+    private Position playerPos;
+
+    /**
+     * Can teleport the player around and between rooms/grids.
+     * @param newGrid The Grid the player should be teleported to
+     * @param playerPosition The position on the new grid where the player has to be teleported.
+     */
+    public Warp(Grid newGrid, Position playerPosition){
+        this.grid = newGrid;
+        this.playerPos = playerPosition;
+    }
+
+    public Grid getGrid() {
+        return this.grid;
+    }
+
+    public Position getPlayerPos() {
+        return this.playerPos;
+    }
+}
